refactor(runtime): extract text_recovery module from agent_loop
The text-based tool-call recovery cluster moves into
`agent_loop/text_recovery.rs`. Three `pub(super)` entry points:

- `looks_like_hallucinated_action` — flags model output that claims a
  completed action (English present-perfect, Italian `ho` + past
  participle, impersonal `è stato/stata`, `successfully X`,
  `messaggio inviato`, …) without an actual tool call. Used by the
  end-turn classifier in `mod.rs` to decide whether to nudge the
  model to retry with a real tool invocation.
- `user_message_has_action_intent` — checks the user's request for
  exact-word action verbs (`send`, `execute`, `create`, `deploy`, …)
  so the hallucination nudge only fires when the user actually asked
  for an action.
- `recover_text_tool_calls` — extracts synthetic `ToolCall` entries
  from free-text LLM output across nine on-the-wire formats:
  `<function=NAME>{…}</function>` (Pattern 1), `<function>NAME{…}</function>`
  (P2), `<tool>NAME{…}</tool>` (P3 — Qwen / DeepSeek), markdown code
  blocks (P4), backtick-wrapped (P5), `[TOOL_CALL]…[/TOOL_CALL]` with
  JSON or arrow syntax (P6 — issue #354), `<tool_call>JSON</tool_call>`
  (P7 — Qwen3, issue #332), `<function name="…" parameters="…" />`
  (P9 — Groq/Llama XML attribute, cached regex per #3491), and the
  bare-JSON last-resort fallback (P8).

Internal helpers (`parse_json_tool_call_object`, `parse_arrow_syntax_tool_call`,
`parse_dash_dash_args`, `try_parse_bare_json_tool_call`) stay private to
the module except for the two that the `mod tests` block exercises
directly (`parse_json_tool_call_object`, `parse_dash_dash_args`) — those
are `pub(super)` and pulled in via the existing test-only import block.

Behavior-preserving: every pattern's matching logic, dedupe key
(`name + input`), recovered-id format (`recovered_<uuid>`), and
warn/info log site are byte-for-byte the same as before the move. The
ordering of pattern-matching (1 → 9 → 8) is unchanged, including the
P8 bare-JSON guard that only runs when `calls.is_empty()`.

mod.rs: 12,641 -> 11,812 LOC (-829). cargo check + cargo clippy --lib
+ cargo check --tests all clean.

Refs #3710
